grouped into component, added separate css and addressed review comments.
