+++ src/styles/vue-styles.scss
@@ -48,6 +48,7 @@
 @import "../ui/toolbar/components/toolbar-checkbox.scss";
 @import "./notebook.scss";
 @import "../plugins/notebook/components/sidebar.scss";
+@import "../utils/imageLargeView/imageLargeView.scss";
 
 #splash-screen {
      display: none;
