--- src/utils/imageLargeView.html
@@ -1,6 +1,0 @@
-<div class="c-notebook-snapshot">
-    <div class="c-notebook-snapshot__header l-browse-bar">
-        <span class="c-object-label__name">{{ time }}</span>
-    </div>
-<img :src="imageSrc" alt="image" :style="filter" />
-</div>
