+++ src/utils/imageLargeView/imageLargeView.html
@@ -1,0 +1,6 @@
+<div class="c-image-large">
+    <div class="c-image-large__header l-browse-bar">
+        <span class="c-object-label__name">{{ time }}</span>
+    </div>
+    <img :src="imageSrc" alt="image" :style="filter" />
+</div>
